Split "Required Hardware" section for readability
The "Required" hardware list was imposing and difficult to parse
before. I'm taking a first stab at making it more approachable by
splitting it into two logical sections: one for the core components of
SecureDrop (the servers) and one for the workflow components (the
workstations, Tails USBs, etc.)

I moved the note about virtualization inside the new subsection on
servers because that seemed like a good place for it.